pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "biopython>=1.85",
    "matplotlib>=3.10.0",
    "nibabel>=5.3.2",
    "nilearn>=0.11.1",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "scipy>=1.15.1",
    "sqlalchemy>=2.0.37",
    "streamlit>=1.42.0",
]
